[Darwin, core] Reorganise labelling of hot/cold partitioning.
Reorganise code to ensure that hot/cold partitioned data ends up in labelled chunks that ld64 can identify as legal atoms.
Although this is for Darwin, it should not adversely affect any other target - in fact, makes the asm slightly more readable.
Also, use section switches in darwin.c (instead of direct calls) for image info - so that we don't lose track of which section we are in.