[Darwin, Core] Implement a hook to allow target-handling of Personality and LSDA indirections.
If an assembler is used that supports .cfi_xxx, then configuration  will detect this and try to use it to generate eh.  This will not work for m32 Darwin since..

The default implementation for indirection will not work for mac-o-pic.  Also the handing of relocations for x86-64 differs for Darwin c.f. other x86-64 targets (which might make assumptions there fragile too).

It might be better to make the hook at a higher level, thus allowing more complete customisation.  This is the minimum.