[Darwin, driver] Back-port: Revise and clean up system version detection.
From: Iain Sandoe <iain@codesourcery.com>
Date: Mon, 7 Sep 2015 09:59:45 +0100

This re-factors the system version detection and makes the version
string available to the darwin_driver_init() routine, when it is
available.  We also delete all the "darwin_minversion" spec stuff
which is unused and redundant.  The default value for compilers is
now set to match the configured default.  If compilers are invoked
directly without an explicit system version, a warning is given.