[Darwin, emutls] Make emutls into a crt
EMUTLS requires a single implementation of the handling of thread-local data (per exe).  So the obvious solution is to make the symbols for this "dynamic lookup" and provide them in the main executable.

However, we might want to build a shared library that could (perhaps) be used with an older GCC (or clang) and thus that isn't sufficient.

So what we now do is to build two crts, one with the symbols "weak".  The weak version is linked into shared libraries and bundles.  The non-weak version is used in main exes.  Therefore, when a shared library is used "stand-alone" without a GCC-linked exe the handling for the shlib will be local.

For this to work, the weak symbols have to be externally-visible on the shared libs.  Thus we have to add them to the libstdc++ list and ensure that they are not made private.