[Darwin, libgcc] Amendments to avoid packaging a shared ext lib.
So now, we've split out the code that needed a shared lib (emutls) and it's possible to default Darwin to link libgcc at the end of the command line (note that there's no need to order it earlier from the point of view of satisfying deps, ld64 will re-visit libs as necessary).

In the case that the implementation of lib routines from libgcc_ext is preferred over the system versions.
(a) include -lgcc on the command line
(b) include -lgcc_ext.1 on the command line and package libgcc_ext.1.dylib with the application.