[Darwin,spec] Update 'as' specs and inputs to handle newer assembler versions.
Based on a patch from Rainer.

A/ Newer versions of ld64 check the min_version command, and newer versions of the system assembler inserts this in response to "-mmacosx-version-min=" on the assembler line.
Unless one makes sensible versions, some object is bound to conflict.

B/ Additionally, there's a difference in behaviour between "as" and "ld" when presented with xx.yy.zz (ld truncates to xx.yy, as doesn't); net result is that one needs to pass a truncated version to "as".

So (if the assembler supports minversion commands)
(a) provide a truncated minversion (as asm_macosx_version_min, which is a driver-only var).
(b) pass this to "as"
(c) Update tests to determine 'HAVE_AS_MMACOSX_VERSION_MIN_OPTION' (Rainer's patch)
(d) For some reason the testcases are "run" (it's not obvious they need to be, they are checking compile-time issues)
  - anyway, to preserve the status quo, I've left them as exec.  However, the minimum version that can be code-gened
   for is target-dependent (there are no x86 versions before 10.4, for example).  To avoid conflicts where the "as" is assuming some minimum, I've set the testversion to 10.5 (which is supported by all the archs we have)
(e) We need to ensure that libgcc and crts are generated with a sufficiently old minversion not to conflict.