[libatomic,configure] Fix saving of CFLAGS
save_CFLAGS is used by other configure macros, which was causing the XCFLAGS set for the first multilib to be present for subsequent ones.
This doesn't show up when the target is 64b with a 32b multilib, but it breaks configuration for a 32b target which supports a 64b multilib.